fix: resolve 5 eval test errors and 2 ESLint errors
- Fix test_runner.py: use dc_store.data.atomic_logs/chains/pending
  instead of dc_store.atomic_logs (attribute moved after DecisionStore
  refactor to Pydantic-based DecisionStoreData)
- Fix ChatInterface ESLint react-refresh/only-export-components:
  extract replaceVariableToken and swapVariableInResult into
  ui/src/utils/variableSwap.ts
